Sketchy-Ass D-Pad Things for the Omniwheels
There wasn't much to do after we got the other code working so I just
made this... thing which works alright (other than diagonals) for moving
the robot according to the dpad. It can't rotate though, and joysticks
are better so we're probably gonna switch this out with Jack's stuff
once he pushes.