VERSION 2 - HUGE CHANGES
We're continuing with the Vuforia Autonomous routine, but we're also
making four others that we can pick between at competition that don't
rely on that: two for each team that can go for beacons with OpenCV or
not use it at all. Major restructuring to the whole system for
Autonomous.